Move 'DownloadAndConfigureMicrosoftPublicKey' to utils
